Extract CFPropertyBundle from Analyzer
Previously, PackageAnalyzer was doing two things:

1. downloading and opening the blob file of an attachment
1. extracting property list from the file using the CFPropertyList library

Only the first part needs to live inside a subclass of
`ActiveStorage::Analyzer` because it's using the `download_blob_to_tempfile`
method. The second part can be extracted into its own library.

The goal is to use this library immediately when a file is uploaded
in the browser as `ActionDispatch::Http::UploadedFile` in order to
extract its version and store it in the database.
